Ported to QT6/KF6. Interim.  Misc cleanup.

+++ src/logic/MP2.h
@@ -30,7 +30,7 @@ class MP2 : public QMediaPlayer
 	Q_OBJECT
 
 	Q_PROPERTY(bool muted READ muted WRITE setMuted NOTIFY mutedChanged)
-	Q_PROPERTY(int volume READ volume WRITE setVolume NOTIFY volumeChanged)
+	Q_PROPERTY(float volume READ volume WRITE setVolume NOTIFY volumeChanged)
 
 public:
 	enum ShuffleSetting {Shuffle, Sequential};
